# storage/api-docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AI4BMI",
        "description": "API E-commerce pour l'app mobile (clients). Authentification : Bearer token (Sanctum) ; seuls POST /register et POST /login sont publics.\n\n**Cycle commande** : création (POST /orders) → statut `pending` → paiement confirmé (webhook ou back-office) → `paid` → expédition → `shipped`. Annulation possible uniquement si `pending` (POST /orders/{id}/cancel). Adresse de livraison : fournie à la commande ou reprise du profil (PATCH /user).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API"
        }
    ],
    "paths": {
...
        "/cart": {
            "get": {
                "tags": [
                    "E-commerce"
                ],
                "summary": "Voir le panier",
                "description": "Panier de l'utilisateur (créé si vide). Lignes avec product et quantity.",
                "operationId": "116cee7f7f059b94f814b380f5c94407",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "E-commerce"
                ],
                "summary": "Vider le panier",
                "description": "Supprime toutes les lignes du panier.",
                "operationId": "7c4dd8337163adfceb3f8eb18e2b5787",
                "responses": {
                    "200": {
                        "description": "Panier vidé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/items": {
            "post": {
                "tags": [
                    "E-commerce"
                ],
                "summary": "Ajouter au panier",
                "description": "Ajoute un produit ou augmente la quantité. Vérification du stock.",
                "operationId": "5a91e43f87df70933884b8d02b9033a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article ajouté"
                    },
                    "422": {
                        "description": "Stock insuffisant"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/items/{cartItem}": {
            "delete": {
                "tags": [
                    "E-commerce"
                ],
                "summary": "Supprimer une ligne du panier",
                "operationId": "85548c85fd2e72000ff4b7187abb1a56",
                "parameters": [
                    {
                        "name": "cartItem",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ligne supprimée"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "E-commerce"
                ],
                "summary": "Modifier une ligne du panier",
                "description": "Mettre à jour la quantité. Si quantity=0, la ligne est supprimée.",
                "operationId": "366322f20dd6e7f97a5ac1e64a6eea22",
                "parameters": [
                    {
                        "name": "cartItem",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}